+++ types/custom-events.d.ts
@@ -1,0 +1,5 @@
+declare namespace CustomEvents {
+    interface FinishedLoading extends CustomEvent {
+        readonly customEvent: 'finishedLoading';
+    }
+}